fix: use fetch mock
