+++ packages/graphql-language-service-server/package.json
@@ -36,14 +36,13 @@
   "peerDependencies": {
     "graphql": "^15.5.0 || ^16.0.0"
   },
-  "COMMENT": "please do not remove depenencies without thorough testing. many dependencies are not imported directly, as they are peer dependencies",
+  "COMMENT": "please do not remove dependencies without thorough testing. many dependencies are not imported directly, as they are peer dependencies",
   "dependencies": {
-    "@astrojs/compiler": "^2.5.0",
     "@babel/parser": "^7.23.6",
     "@babel/types": "^7.23.5",
     "@graphql-tools/code-file-loader": "8.0.3",
     "@vue/compiler-sfc": "^3.4.5",
-    "astrojs-compiler-sync": "^0.3.5",
+    "astrojs-compiler-sync": "1.0.0",
     "cosmiconfig-toml-loader": "^1.0.0",
     "dotenv": "10.0.0",
     "fast-glob": "^3.2.7",
