+++ packages/graphql-language-service-server/src/__tests__/parseDocument-test.ts
@@ -1,7 +1,7 @@
 import { parseDocument } from '../parseDocument';
 
 describe('parseDocument', () => {
-  it('parseDocument finds queries in tagged templates', async () => {
+  it('parseDocument finds queries in tagged templates', () => {
     const text = `
     // @flow
     import {gql} from 'react-apollo';
@@ -32,7 +32,7 @@ describe('parseDocument', () => {
     `);
   });
 
-  it('parseDocument finds queries in tagged templates in leaf', async () => {
+  it('parseDocument finds queries in tagged templates in leaf', () => {
     const text = `
     import {gql} from 'react-apollo';
     import type {B} from 'B';
@@ -58,7 +58,7 @@ describe('parseDocument', () => {
     `);
   });
 
-  it('parseDocument finds queries in tagged templates using typescript', async () => {
+  it('parseDocument finds queries in tagged templates using typescript', () => {
     const text = `
     import {gql} from 'react-apollo';
     import {B} from 'B';
@@ -88,7 +88,7 @@ describe('parseDocument', () => {
     `);
   });
 
-  it('parseDocument finds queries in tagged templates using tsx', async () => {
+  it('parseDocument finds queries in tagged templates using tsx', () => {
     const text = `
     import {gql} from 'react-apollo';
     import {B} from 'B';
@@ -120,7 +120,7 @@ describe('parseDocument', () => {
     `);
   });
 
-  it('parseDocument finds queries in multi-expression tagged templates using tsx', async () => {
+  it('parseDocument finds queries in multi-expression tagged templates using tsx', () => {
     const text = `
     import {gql} from 'react-apollo';
     import {B} from 'B';
@@ -153,7 +153,7 @@ describe('parseDocument', () => {
     }`);
   });
   // TODO: why an extra line here?
-  it('parseDocument finds queries in multi-expression tagged template with declarations with using tsx', async () => {
+  it('parseDocument finds queries in multi-expression tagged template with declarations with using tsx', () => {
     const text = `
     import {gql} from 'react-apollo';
     import {B} from 'B';
@@ -186,7 +186,7 @@ describe('parseDocument', () => {
     }`);
   });
 
-  it('parseDocument finds queries in multi-expression template strings using tsx', async () => {
+  it('parseDocument finds queries in multi-expression template strings using tsx', () => {
     const text = `
     import {gql} from 'react-apollo';
     import {B} from 'B';
@@ -222,7 +222,7 @@ describe('parseDocument', () => {
     `);
   });
 
-  it('parseDocument finds queries in call expressions with template literals', async () => {
+  it('parseDocument finds queries in call expressions with template literals', () => {
     const text = `
     // @flow
     import {gql} from 'react-apollo';
@@ -253,7 +253,7 @@ describe('parseDocument', () => {
     `);
   });
 
-  it('parseDocument finds queries in #graphql-annotated templates', async () => {
+  it('parseDocument finds queries in #graphql-annotated templates', () => {
     const text = `
     import {gql} from 'react-apollo';
     import {B} from 'B';
@@ -283,7 +283,7 @@ describe('parseDocument', () => {
     `);
   });
 
-  it('parseDocument finds queries in /*GraphQL*/-annotated templates', async () => {
+  it('parseDocument finds queries in /*GraphQL*/-annotated templates', () => {
     const text = `
     import {gql} from 'react-apollo';
     import {B} from 'B';
@@ -314,7 +314,7 @@ describe('parseDocument', () => {
     `);
   });
 
-  it('parseDocument ignores non gql tagged templates', async () => {
+  it('parseDocument ignores non gql tagged templates', () => {
     const text = `
     // @flow
     import randomThing from 'package';
@@ -337,7 +337,7 @@ describe('parseDocument', () => {
     expect(contents.length).toEqual(0);
   });
 
-  it('parseDocument ignores non gql call expressions with template literals', async () => {
+  it('parseDocument ignores non gql call expressions with template literals', () => {
     const text = `
     // @flow
     import randomthing from 'package';
@@ -360,7 +360,7 @@ describe('parseDocument', () => {
     expect(contents.length).toEqual(0);
   });
 
-  it('an unparsable JS/TS file does not throw and bring down the server', async () => {
+  it('an unparsable JS/TS file does not throw and bring down the server', () => {
     const text = `
     // @flow
     import type randomThing fro 'package';
@@ -380,14 +380,14 @@ describe('parseDocument', () => {
     expect(contents.length).toEqual(0);
   });
 
-  it('an empty file is ignored', async () => {
+  it('an empty file is ignored', () => {
     const text = '';
 
     const contents = parseDocument(text, 'test.js');
     expect(contents.length).toEqual(0);
   });
 
-  it('a whitespace only file with empty asts is ignored', async () => {
+  it('a whitespace only file with empty asts is ignored', () => {
     const text = `
     
     `;
@@ -396,7 +396,7 @@ describe('parseDocument', () => {
     expect(contents.length).toEqual(0);
   });
 
-  it('an ignored file is ignored', async () => {
+  it('an ignored file is ignored', () => {
     const text = `
     something
     `;
