+++ packages/graphql-language-service-server/src/startServer.ts
@@ -7,7 +7,7 @@
  *
  */
 import { MessageProcessor } from './MessageProcessor';
-import { GraphQLConfig, GraphQLExtensionDeclaration } from 'graphql-config';
+import { GraphQLConfig } from 'graphql-config';
 import {
   IPCMessageReader,
   IPCMessageWriter,
@@ -46,79 +46,8 @@ import {
   DEFAULT_SUPPORTED_GRAPHQL_EXTENSIONS,
   SupportedExtensionsEnum,
 } from './constants';
-import { LoadConfigOptions } from './types';
+import { LoadConfigOptions, ServerOptions } from './types';
 import { createConnection } from 'node:net';
-
-export interface ServerOptions {
-  /**
-   * socket, streams, or node (ipc).
-   * @default 'node'
-   */
-  method?: 'socket' | 'stream' | 'node';
-  /**
-   * (socket only) port for the LSP server to run on. required if using method socket
-   */
-  port?: number;
-  /**
-   * (socket only) hostname for the LSP server to run on.
-   * @default '127.0.0.1'
-   */
-  hostname?: string;
-  /**
-   * (socket only) encoding for the LSP server to use.
-   * @default 'utf-8'
-   */
-  encoding?: 'utf-8' | 'ascii';
-  /**
-   * `LoadConfigOptions` from `graphql-config@3` to use when we `loadConfig()`
-   * uses process.cwd() by default for `rootDir` option.
-   * you can also pass explicit `filepath`, add extensions, etc
-   */
-  loadConfigOptions?: LoadConfigOptions;
-  /**
-   * @deprecated use loadConfigOptions.rootDir now) the directory where graphql-config is found
-   */
-  configDir?: string;
-  /**
-   * @deprecated use loadConfigOptions.extensions
-   */
-  extensions?: GraphQLExtensionDeclaration[];
-  /**
-   * allowed file extensions for embedded graphql, used by the parser.
-   * note that with vscode, this is also controlled by manifest and client configurations.
-   * do not put full-file graphql extensions here!
-   * @default ['.js', '.jsx', '.tsx', '.ts', '.mjs']
-   */
-  fileExtensions?: ReadonlyArray<SupportedExtensionsEnum>;
-  /**
-   * allowed file extensions for full-file graphql, used by the parser.
-   * @default ['graphql', 'graphqls', 'gql'  ]
-   */
-  graphqlFileExtensions?: string[];
-  /**
-   * pre-existing GraphQLConfig primitive, to override `loadConfigOptions` and related deprecated fields
-   */
-  config?: GraphQLConfig;
-  /**
-   * custom, multi-language parser used by the LSP server.
-   * detects extension from uri and decides how to parse it.
-   * uses graphql.parse() by default
-   * response format is designed to assist with developing LSP tooling around embedded language support
-   */
-  parser?: typeof parseDocument;
-  /**
-   * the temporary directory that the server writes to for logs and caching schema
-   */
-  tmpDir?: string;
-
-  /**
-   * debug mode
-   *
-   * same as with the client reference implementation, the debug setting controls logging output
-   * this allows all logger.info() messages to come through. by default, the highest level is warn
-   */
-  debug?: true;
-}
 
 /**
  * Make loadConfigOptions
@@ -163,26 +92,23 @@ export const buildOptions = (options: ServerOptions): MappedServerOptions => {
  */
 export default async function startServer(
   options?: ServerOptions,
-): Promise<void> {
-  if (!options?.method) {
-    return;
-  }
-  const finalOptions = buildOptions(options);
+): Promise<Connection | void> {
+  const finalOptions = buildOptions({ method: 'node', ...options });
   let reader;
   let writer;
-  switch (options.method) {
+  switch (finalOptions.method) {
     case 'socket':
       // For socket connection, the message connection needs to be
       // established before the server socket starts listening.
       // Do that, and return at the end of this block.
-      if (!options.port) {
+      if (!finalOptions.port) {
         process.stderr.write(
           '--port is required to establish socket connection.',
         );
         process.exit(1);
       }
 
-      const { port, hostname, encoding } = options;
+      const { port, hostname, encoding } = finalOptions;
       const socket = createConnection(port, hostname ?? '127.0.01');
 
       reader = new SocketMessageReader(socket, encoding ?? 'utf-8');
@@ -190,13 +116,20 @@ export default async function startServer(
 
       break;
     case 'stream':
-      reader = new StreamMessageReader(process.stdin);
-      writer = new StreamMessageWriter(process.stdout);
-      break;
+      const server = createLanguageServerConnection(
+        // @ts-expect-error this still works, just a type mismatch
+        process.stdin,
+        process.stderr,
+        {
+          connectionStrategy: 'stdio',
+        },
+      );
+      server.listen();
+      return server;
+
     default:
       reader = new IPCMessageReader(process);
       writer = new IPCMessageWriter(process);
-
       break;
   }
   const streamServer = await initializeHandlers({
@@ -205,6 +138,7 @@ export default async function startServer(
     options: finalOptions,
   });
   streamServer.listen();
+  return streamServer;
 }
 
 type InitializerParams = {
