+++ packages/graphql-language-service/src/parser/index.ts
@@ -15,4 +15,15 @@ export { butNot, list, opt, p, t } from './RuleHelpers';
 
 export { default as onlineParser, ParserOptions } from './onlineParser';
 
+export {
+  runOnlineParser,
+  type ParserCallbackFn,
+  getTokenAtPosition,
+  getContextAtPosition,
+  GraphQLDocumentMode,
+  getDocumentMode,
+} from './api';
+
+export { getTypeInfo, getDefinitionState, getFieldDef } from './getTypeInfo';
+
 export * from './types';
